chess-engine/requirements.txt
connexion==2.2.0
Flask==1.0.2
gunicorn==19.9.0
gym==0.10.5
numpy==1.14.5
pandas==0.23.3
Keras==2.2.4
python-chess==0.25.0
Werkzeug==0.14.1

